Move to v0.0.5
