--- src/modules/0.0.4/aws_ecs_quickstart/module.json
@@ -1,5 +1,0 @@
-{
-  "name": "aws_ecs_quickstart",
-  "version": "0.0.4",
-  "dependencies": ["iasql_platform@0.0.4", "aws_account@0.0.4"]
-}
