+++ src/modules/0.0.4/index.ts
@@ -5,7 +5,6 @@ export * from './aws_cloudwatch'
 export * from './aws_ec2'
 export * from './aws_ecr'
 export * from './aws_ecs_fargate'
-export * from './aws_ecs_quickstart'
 export * from './aws_elb'
 export * from './aws_rds'
 export * from './aws_route53_hosted_zones'
