+++ src/modules/0.0.5/aws_ecs_quickstart/entity/ecs_quickstart.ts
@@ -7,7 +7,7 @@ import {
   PrimaryColumn,
   Check,
 } from 'typeorm'
-import { CpuMemCombination } from '../../aws_ecs_fargate/entity';
+import { CpuMemCombination } from '../../../0.0.4/aws_ecs_fargate/entity';
 
 @Check(`("image_tag" is null and "image_digest" is null) or ("image_tag" is not null and "image_digest" is null) or ("image_tag" is null and "image_digest" is not null)`)
 @Entity()
