+++ src/modules/0.0.5/aws_ecs_quickstart/entity/index.ts
