+++ src/modules/0.0.5/aws_ecs_quickstart/index.ts
@@ -3,12 +3,12 @@ import logger, { debugObj } from '../../../services/logger'
 import { EcsQuickstart } from './entity'
 import { Context, Crud, Mapper, Module, } from '../../interfaces'
 import * as metadata from './module.json'
-import { SecurityGroup, SecurityGroupRule } from '../aws_security_group/entity'
-import { ActionTypeEnum, IpAddressType, Listener, LoadBalancer, LoadBalancerSchemeEnum, LoadBalancerTypeEnum, ProtocolEnum, TargetGroup, TargetTypeEnum } from '../aws_elb/entity'
-import { LogGroup } from '../aws_cloudwatch/entity'
-import { ImageTagMutability, Repository } from '../aws_ecr/entity'
-import { Role } from '../aws_iam/entity'
-import { AssignPublicIp, Cluster, ContainerDefinition, CpuMemCombination, Service, TaskDefinition, TransportProtocol } from '../aws_ecs_fargate/entity'
+import { SecurityGroup, SecurityGroupRule } from '../../0.0.4/aws_security_group/entity'
+import { ActionTypeEnum, IpAddressType, Listener, LoadBalancer, LoadBalancerSchemeEnum, LoadBalancerTypeEnum, ProtocolEnum, TargetGroup, TargetTypeEnum } from '../../0.0.4/aws_elb/entity'
+import { LogGroup } from '../../0.0.4/aws_cloudwatch/entity'
+import { ImageTagMutability, Repository } from '../../0.0.4/aws_ecr/entity'
+import { Role } from '../../0.0.4/aws_iam/entity'
+import { AssignPublicIp, Cluster, ContainerDefinition, CpuMemCombination, Service, TaskDefinition, TransportProtocol } from '../../0.0.4/aws_ecs_fargate/entity'
 import { Subnet, Vpc } from '@aws-sdk/client-ec2'
 import { CreateLoadBalancerCommandInput } from '@aws-sdk/client-elastic-load-balancing-v2'
 import { Service as AwsService } from '@aws-sdk/client-ecs'
