+++ src/modules/0.0.5/aws_ecs_quickstart/module.json
@@ -1,0 +1,5 @@
+{
+  "name": "aws_ecs_quickstart",
+  "version": "0.0.5",
+  "dependencies": ["iasql_platform@0.0.5", "aws_account@0.0.5"]
+}
