+++ src/modules/0.0.5/index.ts
@@ -5,6 +5,7 @@ export * from './aws_cloudwatch'
 export * from './aws_ec2'
 export * from './aws_ecr'
 export * from './aws_ecs_fargate'
+export * from './aws_ecs_quickstart'
 export * from './aws_elb'
 export * from './aws_rds'
 export * from './aws_route53_hosted_zones'
